+++ src/pkcs1v15/decrypting_key.rs
@@ -1,0 +1,51 @@
+use super::{decrypt, EncryptingKey};
+use crate::{
+    dummy_rng::DummyRng,
+    traits::{Decryptor, EncryptingKeypair, RandomizedDecryptor},
+    Result, RsaPrivateKey,
+};
+use alloc::vec::Vec;
+use rand_core::CryptoRngCore;
+use zeroize::ZeroizeOnDrop;
+
+/// Decryption key for PKCS#1 v1.5 decryption as described in [RFC8017 § 7.2].
+///
+/// [RFC8017 § 7.2]: https://datatracker.ietf.org/doc/html/rfc8017#section-7.2
+#[derive(Debug, Clone)]
+pub struct DecryptingKey {
+    inner: RsaPrivateKey,
+}
+
+impl DecryptingKey {
+    /// Create a new verifying key from an RSA public key.
+    pub fn new(key: RsaPrivateKey) -> Self {
+        Self { inner: key }
+    }
+}
+
+impl Decryptor for DecryptingKey {
+    fn decrypt(&self, ciphertext: &[u8]) -> Result<Vec<u8>> {
+        decrypt::<DummyRng>(None, &self.inner, ciphertext)
+    }
+}
+
+impl RandomizedDecryptor for DecryptingKey {
+    fn decrypt_with_rng<R: CryptoRngCore + ?Sized>(
+        &self,
+        rng: &mut R,
+        ciphertext: &[u8],
+    ) -> Result<Vec<u8>> {
+        decrypt(Some(rng), &self.inner, ciphertext)
+    }
+}
+
+impl EncryptingKeypair for DecryptingKey {
+    type EncryptingKey = EncryptingKey;
+    fn encrypting_key(&self) -> EncryptingKey {
+        EncryptingKey {
+            inner: self.inner.clone().into(),
+        }
+    }
+}
+
+impl ZeroizeOnDrop for DecryptingKey {}
