+++ src/pss.rs
@@ -9,39 +9,33 @@
 //! [Probabilistic Signature Scheme]: https://en.wikipedia.org/wiki/Probabilistic_signature_scheme
 //! [RFC8017 § 8.1]: https://datatracker.ietf.org/doc/html/rfc8017#section-8.1
 
-use alloc::{boxed::Box, string::ToString, vec::Vec};
-use core::fmt::{self, Debug, Display, Formatter, LowerHex, UpperHex};
-use core::marker::PhantomData;
+mod blinded_signing_key;
+mod signature;
+mod signing_key;
+mod verifying_key;
+
+pub use self::{
+    blinded_signing_key::BlindedSigningKey, signature::Signature, signing_key::SigningKey,
+    verifying_key::VerifyingKey,
+};
+
+use alloc::{boxed::Box, vec::Vec};
+use core::fmt::{self, Debug};
 
 use const_oid::{AssociatedOid, ObjectIdentifier};
 use digest::{Digest, DynDigest, FixedOutputReset};
 use num_bigint::BigUint;
 use pkcs1::RsaPssParams;
-use pkcs8::{
-    spki::{
-        der::{Any, AnyRef},
-        AlgorithmIdentifierOwned, AlgorithmIdentifierRef, AssociatedAlgorithmIdentifier,
-        DynSignatureAlgorithmIdentifier,
-    },
-    Document, EncodePrivateKey, EncodePublicKey, SecretDocument,
-};
+use pkcs8::spki::{der::Any, AlgorithmIdentifierOwned};
 use rand_core::CryptoRngCore;
-use signature::{
-    hazmat::{PrehashVerifier, RandomizedPrehashSigner},
-    DigestVerifier, Keypair, RandomizedDigestSigner, RandomizedSigner, SignatureEncoding, Verifier,
-};
-use zeroize::ZeroizeOnDrop;
 
 use crate::algorithms::pad::{uint_to_be_pad, uint_to_zeroizing_be_pad};
 use crate::algorithms::pss::*;
 use crate::algorithms::rsa::{rsa_decrypt_and_check, rsa_encrypt};
 use crate::errors::{Error, Result};
-use crate::traits::SignatureScheme;
 use crate::traits::PublicKeyParts;
+use crate::traits::SignatureScheme;
 use crate::{RsaPrivateKey, RsaPublicKey};
-
-#[cfg(feature = "getrandom")]
-use {rand_core::OsRng, signature::Signer};
 
 /// Digital signatures using PSS padding.
 pub struct Pss {
@@ -126,62 +120,6 @@ impl Debug for Pss {
             .field("digest", &"...")
             .field("salt_len", &self.salt_len)
             .finish()
-    }
-}
-
-/// RSASSA-PSS signatures as described in [RFC8017 § 8.1].
-///
-/// [RFC8017 § 8.1]: https://datatracker.ietf.org/doc/html/rfc8017#section-8.1
-#[derive(Clone, PartialEq, Eq)]
-pub struct Signature {
-    inner: BigUint,
-    len: usize,
-}
-
-impl SignatureEncoding for Signature {
-    type Repr = Box<[u8]>;
-}
-
-impl TryFrom<&[u8]> for Signature {
-    type Error = signature::Error;
-
-    fn try_from(bytes: &[u8]) -> signature::Result<Self> {
-        Ok(Self {
-            len: bytes.len(),
-            inner: BigUint::from_bytes_be(bytes),
-        })
-    }
-}
-
-impl From<Signature> for Box<[u8]> {
-    fn from(signature: Signature) -> Box<[u8]> {
-        signature.inner.to_bytes_be().into_boxed_slice()
-    }
-}
-
-impl Debug for Signature {
-    fn fmt(&self, fmt: &mut Formatter<'_>) -> core::result::Result<(), core::fmt::Error> {
-        fmt.debug_tuple("Signature")
-            .field(&self.to_string())
-            .finish()
-    }
-}
-
-impl LowerHex for Signature {
-    fn fmt(&self, f: &mut Formatter<'_>) -> core::fmt::Result {
-        write!(f, "{:x}", &self.inner)
-    }
-}
-
-impl UpperHex for Signature {
-    fn fmt(&self, f: &mut Formatter<'_>) -> core::fmt::Result {
-        write!(f, "{:X}", &self.inner)
-    }
-}
-
-impl Display for Signature {
-    fn fmt(&self, f: &mut Formatter<'_>) -> core::fmt::Result {
-        write!(f, "{:X}", self)
     }
 }
 
@@ -289,64 +227,6 @@ fn sign_pss_with_salt_digest<T: CryptoRngCore + ?Sized, D: Digest + FixedOutputR
     )
 }
 
-/// Signing key for producing RSASSA-PSS signatures as described in
-/// [RFC8017 § 8.1].
-///
-/// [RFC8017 § 8.1]: https://datatracker.ietf.org/doc/html/rfc8017#section-8.1
-#[derive(Debug, Clone)]
-pub struct SigningKey<D>
-where
-    D: Digest,
-{
-    inner: RsaPrivateKey,
-    salt_len: usize,
-    phantom: PhantomData<D>,
-}
-
-impl<D> SigningKey<D>
-where
-    D: Digest,
-{
-    /// Create a new RSASSA-PSS signing key.
-    /// Digest output size is used as a salt length.
-    pub fn new(key: RsaPrivateKey) -> Self {
-        Self::new_with_salt_len(key, <D as Digest>::output_size())
-    }
-
-    /// Create a new RSASSA-PSS signing key with a salt of the given length.
-    pub fn new_with_salt_len(key: RsaPrivateKey, salt_len: usize) -> Self {
-        Self {
-            inner: key,
-            salt_len,
-            phantom: Default::default(),
-        }
-    }
-
-    /// Generate a new random RSASSA-PSS signing key.
-    /// Digest output size is used as a salt length.
-    pub fn random<R: CryptoRngCore + ?Sized>(rng: &mut R, bit_size: usize) -> Result<Self> {
-        Self::random_with_salt_len(rng, bit_size, <D as Digest>::output_size())
-    }
-
-    /// Generate a new random RSASSA-PSS signing key with a salt of the given length.
-    pub fn random_with_salt_len<R: CryptoRngCore + ?Sized>(
-        rng: &mut R,
-        bit_size: usize,
-        salt_len: usize,
-    ) -> Result<Self> {
-        Ok(Self {
-            inner: RsaPrivateKey::new(rng, bit_size)?,
-            salt_len,
-            phantom: Default::default(),
-        })
-    }
-
-    /// Return specified salt length for this key
-    pub fn salt_len(&self) -> usize {
-        self.salt_len
-    }
-}
-
 fn get_pss_signature_algo_id<D>(salt_len: u8) -> pkcs8::spki::Result<AlgorithmIdentifierOwned>
 where
     D: Digest + AssociatedOid,
@@ -359,447 +239,6 @@ where
         oid: ID_RSASSA_PSS,
         parameters: Some(Any::encode_from(&pss_params)?),
     })
-}
-
-impl<D> AssociatedAlgorithmIdentifier for SigningKey<D>
-where
-    D: Digest,
-{
-    type Params = AnyRef<'static>;
-
-    const ALGORITHM_IDENTIFIER: AlgorithmIdentifierRef<'static> = pkcs1::ALGORITHM_ID;
-}
-
-impl<D> DynSignatureAlgorithmIdentifier for SigningKey<D>
-where
-    D: Digest + AssociatedOid,
-{
-    fn signature_algorithm_identifier(&self) -> pkcs8::spki::Result<AlgorithmIdentifierOwned> {
-        get_pss_signature_algo_id::<D>(self.salt_len as u8)
-    }
-}
-
-impl<D> From<RsaPrivateKey> for SigningKey<D>
-where
-    D: Digest,
-{
-    fn from(key: RsaPrivateKey) -> Self {
-        Self::new(key)
-    }
-}
-
-impl<D> From<SigningKey<D>> for RsaPrivateKey
-where
-    D: Digest,
-{
-    fn from(key: SigningKey<D>) -> Self {
-        key.inner
-    }
-}
-
-impl<D> EncodePrivateKey for SigningKey<D>
-where
-    D: Digest,
-{
-    fn to_pkcs8_der(&self) -> pkcs8::Result<SecretDocument> {
-        self.inner.to_pkcs8_der()
-    }
-}
-
-impl<D> Keypair for SigningKey<D>
-where
-    D: Digest,
-{
-    type VerifyingKey = VerifyingKey<D>;
-    fn verifying_key(&self) -> Self::VerifyingKey {
-        VerifyingKey {
-            inner: self.inner.to_public_key(),
-            salt_len: self.salt_len,
-            phantom: Default::default(),
-        }
-    }
-}
-
-#[cfg(feature = "getrandom")]
-impl<D> Signer<Signature> for SigningKey<D>
-where
-    D: Digest + FixedOutputReset,
-{
-    fn try_sign(&self, msg: &[u8]) -> signature::Result<Signature> {
-        self.try_sign_with_rng(&mut OsRng, msg)
-    }
-}
-
-impl<D> RandomizedSigner<Signature> for SigningKey<D>
-where
-    D: Digest + FixedOutputReset,
-{
-    fn try_sign_with_rng(
-        &self,
-        rng: &mut impl CryptoRngCore,
-        msg: &[u8],
-    ) -> signature::Result<Signature> {
-        sign_digest::<_, D>(rng, false, &self.inner, &D::digest(msg), self.salt_len)?
-            .as_slice()
-            .try_into()
-    }
-}
-
-impl<D> RandomizedDigestSigner<D, Signature> for SigningKey<D>
-where
-    D: Digest + FixedOutputReset,
-{
-    fn try_sign_digest_with_rng(
-        &self,
-        rng: &mut impl CryptoRngCore,
-        digest: D,
-    ) -> signature::Result<Signature> {
-        sign_digest::<_, D>(rng, false, &self.inner, &digest.finalize(), self.salt_len)?
-            .as_slice()
-            .try_into()
-    }
-}
-
-impl<D> RandomizedPrehashSigner<Signature> for SigningKey<D>
-where
-    D: Digest + FixedOutputReset,
-{
-    fn sign_prehash_with_rng(
-        &self,
-        rng: &mut impl CryptoRngCore,
-        prehash: &[u8],
-    ) -> signature::Result<Signature> {
-        sign_digest::<_, D>(rng, false, &self.inner, prehash, self.salt_len)?
-            .as_slice()
-            .try_into()
-    }
-}
-
-impl<D> AsRef<RsaPrivateKey> for SigningKey<D>
-where
-    D: Digest,
-{
-    fn as_ref(&self) -> &RsaPrivateKey {
-        &self.inner
-    }
-}
-
-impl<D> ZeroizeOnDrop for SigningKey<D> where D: Digest {}
-
-/// Signing key for producing "blinded" RSASSA-PSS signatures as described in
-/// [draft-irtf-cfrg-rsa-blind-signatures](https://datatracker.ietf.org/doc/draft-irtf-cfrg-rsa-blind-signatures/).
-#[derive(Debug, Clone)]
-pub struct BlindedSigningKey<D>
-where
-    D: Digest,
-{
-    inner: RsaPrivateKey,
-    salt_len: usize,
-    phantom: PhantomData<D>,
-}
-
-impl<D> BlindedSigningKey<D>
-where
-    D: Digest,
-{
-    /// Create a new RSASSA-PSS signing key which produces "blinded"
-    /// signatures.
-    /// Digest output size is used as a salt length.
-    pub fn new(key: RsaPrivateKey) -> Self {
-        Self::new_with_salt_len(key, <D as Digest>::output_size())
-    }
-
-    /// Create a new RSASSA-PSS signing key which produces "blinded"
-    /// signatures with a salt of the given length.
-    pub fn new_with_salt_len(key: RsaPrivateKey, salt_len: usize) -> Self {
-        Self {
-            inner: key,
-            salt_len,
-            phantom: Default::default(),
-        }
-    }
-
-    /// Create a new random RSASSA-PSS signing key which produces "blinded"
-    /// signatures.
-    /// Digest output size is used as a salt length.
-    pub fn random<R: CryptoRngCore + ?Sized>(rng: &mut R, bit_size: usize) -> Result<Self> {
-        Self::random_with_salt_len(rng, bit_size, <D as Digest>::output_size())
-    }
-
-    /// Create a new random RSASSA-PSS signing key which produces "blinded"
-    /// signatures with a salt of the given length.
-    pub fn random_with_salt_len<R: CryptoRngCore + ?Sized>(
-        rng: &mut R,
-        bit_size: usize,
-        salt_len: usize,
-    ) -> Result<Self> {
-        Ok(Self {
-            inner: RsaPrivateKey::new(rng, bit_size)?,
-            salt_len,
-            phantom: Default::default(),
-        })
-    }
-
-    /// Return specified salt length for this key
-    pub fn salt_len(&self) -> usize {
-        self.salt_len
-    }
-}
-
-impl<D> AssociatedAlgorithmIdentifier for BlindedSigningKey<D>
-where
-    D: Digest,
-{
-    type Params = AnyRef<'static>;
-
-    const ALGORITHM_IDENTIFIER: AlgorithmIdentifierRef<'static> = pkcs1::ALGORITHM_ID;
-}
-
-impl<D> DynSignatureAlgorithmIdentifier for BlindedSigningKey<D>
-where
-    D: Digest + AssociatedOid,
-{
-    fn signature_algorithm_identifier(&self) -> pkcs8::spki::Result<AlgorithmIdentifierOwned> {
-        get_pss_signature_algo_id::<D>(self.salt_len as u8)
-    }
-}
-
-impl<D> From<RsaPrivateKey> for BlindedSigningKey<D>
-where
-    D: Digest,
-{
-    fn from(key: RsaPrivateKey) -> Self {
-        Self::new(key)
-    }
-}
-
-impl<D> From<BlindedSigningKey<D>> for RsaPrivateKey
-where
-    D: Digest,
-{
-    fn from(key: BlindedSigningKey<D>) -> Self {
-        key.inner
-    }
-}
-
-impl<D> EncodePrivateKey for BlindedSigningKey<D>
-where
-    D: Digest,
-{
-    fn to_pkcs8_der(&self) -> pkcs8::Result<SecretDocument> {
-        self.inner.to_pkcs8_der()
-    }
-}
-
-impl<D> Keypair for BlindedSigningKey<D>
-where
-    D: Digest,
-{
-    type VerifyingKey = VerifyingKey<D>;
-    fn verifying_key(&self) -> Self::VerifyingKey {
-        VerifyingKey {
-            inner: self.inner.to_public_key(),
-            salt_len: self.salt_len,
-            phantom: Default::default(),
-        }
-    }
-}
-
-impl<D> RandomizedSigner<Signature> for BlindedSigningKey<D>
-where
-    D: Digest + FixedOutputReset,
-{
-    fn try_sign_with_rng(
-        &self,
-        rng: &mut impl CryptoRngCore,
-        msg: &[u8],
-    ) -> signature::Result<Signature> {
-        sign_digest::<_, D>(rng, true, &self.inner, &D::digest(msg), self.salt_len)?
-            .as_slice()
-            .try_into()
-    }
-}
-
-impl<D> RandomizedDigestSigner<D, Signature> for BlindedSigningKey<D>
-where
-    D: Digest + FixedOutputReset,
-{
-    fn try_sign_digest_with_rng(
-        &self,
-        rng: &mut impl CryptoRngCore,
-        digest: D,
-    ) -> signature::Result<Signature> {
-        sign_digest::<_, D>(rng, true, &self.inner, &digest.finalize(), self.salt_len)?
-            .as_slice()
-            .try_into()
-    }
-}
-
-impl<D> RandomizedPrehashSigner<Signature> for BlindedSigningKey<D>
-where
-    D: Digest + FixedOutputReset,
-{
-    fn sign_prehash_with_rng(
-        &self,
-        rng: &mut impl CryptoRngCore,
-        prehash: &[u8],
-    ) -> signature::Result<Signature> {
-        sign_digest::<_, D>(rng, true, &self.inner, prehash, self.salt_len)?
-            .as_slice()
-            .try_into()
-    }
-}
-
-impl<D> AsRef<RsaPrivateKey> for BlindedSigningKey<D>
-where
-    D: Digest,
-{
-    fn as_ref(&self) -> &RsaPrivateKey {
-        &self.inner
-    }
-}
-
-impl<D> ZeroizeOnDrop for BlindedSigningKey<D> where D: Digest {}
-
-/// Verifying key for checking the validity of RSASSA-PSS signatures as
-/// described in [RFC8017 § 8.1].
-///
-/// [RFC8017 § 8.1]: https://datatracker.ietf.org/doc/html/rfc8017#section-8.1
-#[derive(Debug)]
-pub struct VerifyingKey<D>
-where
-    D: Digest,
-{
-    inner: RsaPublicKey,
-    salt_len: usize,
-    phantom: PhantomData<D>,
-}
-
-/* Implemented manually so we don't have to bind D with Clone */
-impl<D> Clone for VerifyingKey<D>
-where
-    D: Digest,
-{
-    fn clone(&self) -> Self {
-        Self {
-            inner: self.inner.clone(),
-            salt_len: self.salt_len,
-            phantom: Default::default(),
-        }
-    }
-}
-
-impl<D> VerifyingKey<D>
-where
-    D: Digest,
-{
-    /// Create a new RSASSA-PSS verifying key.
-    /// Digest output size is used as a salt length.
-    pub fn new(key: RsaPublicKey) -> Self {
-        Self::new_with_salt_len(key, <D as Digest>::output_size())
-    }
-
-    /// Create a new RSASSA-PSS verifying key.
-    pub fn new_with_salt_len(key: RsaPublicKey, salt_len: usize) -> Self {
-        Self {
-            inner: key,
-            salt_len,
-            phantom: Default::default(),
-        }
-    }
-}
-
-impl<D> AssociatedAlgorithmIdentifier for VerifyingKey<D>
-where
-    D: Digest,
-{
-    type Params = AnyRef<'static>;
-
-    const ALGORITHM_IDENTIFIER: AlgorithmIdentifierRef<'static> = pkcs1::ALGORITHM_ID;
-}
-
-impl<D> From<RsaPublicKey> for VerifyingKey<D>
-where
-    D: Digest,
-{
-    fn from(key: RsaPublicKey) -> Self {
-        Self::new(key)
-    }
-}
-
-impl<D> From<VerifyingKey<D>> for RsaPublicKey
-where
-    D: Digest,
-{
-    fn from(key: VerifyingKey<D>) -> Self {
-        key.inner
-    }
-}
-
-impl<D> Verifier<Signature> for VerifyingKey<D>
-where
-    D: Digest + FixedOutputReset,
-{
-    fn verify(&self, msg: &[u8], signature: &Signature) -> signature::Result<()> {
-        verify_digest::<D>(
-            &self.inner,
-            &D::digest(msg),
-            &signature.inner,
-            signature.len,
-            self.salt_len,
-        )
-        .map_err(|e| e.into())
-    }
-}
-
-impl<D> DigestVerifier<D, Signature> for VerifyingKey<D>
-where
-    D: Digest + FixedOutputReset,
-{
-    fn verify_digest(&self, digest: D, signature: &Signature) -> signature::Result<()> {
-        verify_digest::<D>(
-            &self.inner,
-            &digest.finalize(),
-            &signature.inner,
-            signature.len,
-            self.salt_len,
-        )
-        .map_err(|e| e.into())
-    }
-}
-
-impl<D> PrehashVerifier<Signature> for VerifyingKey<D>
-where
-    D: Digest + FixedOutputReset,
-{
-    fn verify_prehash(&self, prehash: &[u8], signature: &Signature) -> signature::Result<()> {
-        verify_digest::<D>(
-            &self.inner,
-            prehash,
-            &signature.inner,
-            signature.len,
-            self.salt_len,
-        )
-        .map_err(|e| e.into())
-    }
-}
-
-impl<D> AsRef<RsaPublicKey> for VerifyingKey<D>
-where
-    D: Digest,
-{
-    fn as_ref(&self) -> &RsaPublicKey {
-        &self.inner
-    }
-}
-
-impl<D> EncodePublicKey for VerifyingKey<D>
-where
-    D: Digest,
-{
-    fn to_public_key_der(&self) -> pkcs8::spki::Result<Document> {
-        self.inner.to_public_key_der()
-    }
 }
 
 #[cfg(test)]
